fix: rename components sidebar.tsx
